+++ ffc-ahwr-event/azure-storage/create-table-client.mjs
